images/boot-infra: actually add the boot infra
adapted from the pixiecore work. this image serves two HTTP endpoints.

/boot.ipxe?mac=&serial= is chained from initial DHCP netboot reply. it
attempts to lookup a k8s Node with given mac or serial, and falls back
to creating a node with a generated petname otherwise.

/ignition will serve up the Ignition config for a Node, it's assumed
there's a "base.ign" config, and other profiles are merged in based on
Node labels/annotations. The hostname is also set based on the Node name

Signed-off-by: Sam Day <me@samcday.com>