hub/cloud-cluster: manually set externalIP for the moment
the advertise-address job is failing, because this control plane is now
using a ClusterIP Service for the apiserver. I should fix that for
robustness reasons, but the issue is right now there's no proper way for
the chart to "know" the external (tailnet) IP is, since we're managing
that separately. The reason we're doing *that* is because headscale
doesn't support OAuth or the Tailscale.com API, and thus the operator
isn't available.

If it were, I believe it already updates the
.status.loadBalancer.ingress[].ip field and thus the existing Job would
work as expected.