hub/cloud-cluster: temporary fix for control plane access
Pods are struggling to resolve the tailnet.hub.samcday.com addresses
since they're only resolveable via Tailscale's MagicDNS
(100.100.100.100) endpoint. There might be a way to fix that properly
but I vaguely recall bashing my head against this in the past and am
assuredly not in the market for (additional) brain damage right now.

So for now we just connect directly the tailnet IP. If that thing
changes the fragile hack I have in place on the apiserver side will
also break. So.