treewide: DRY up Butane configs
Introduce common/butane and move a bunch of stuff here.

Butane's config merging is used to pull these common bits in where
needed.

Update bootie deployments and hub's bootstrap.sh to compile the
common configs and make those available for the config merging.

For now the configs in common/butane are not encrypted since I couldn't
think of any reasons they should be. Maybe that'll change in future.